Duplicates

+++ src/main/kotlin/api/Currency.kt
@@ -2,6 +2,6 @@ package api
 
 data class Currency (
 	val code : String,
-	val rate : Int,
+	val rate : Double,
 	val fixed : Boolean
 )
